+++ CHANGELOG.md
@@ -23,7 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
--
+- Don't display the arg `--upmix-50-to-51` for **DDP-BluRay**.
 
 ### Fixed